add likes and comments routes

+++ src/db/models/comment.js
@@ -1,7 +1,7 @@
 export default {
   CREATE_TABLE: `CREATE TABLE IF NOT EXISTS comments (
     id serial PRIMARY KEY,
-    post_id int , 
+    post_id NUMERIC , 
     author_id VARCHAR (50) ,
     content text,
     FOREIGN KEY (post_id) REFERENCES posts(id) ON UPDATE CASCADE,
